refactor(step4): Abstract 클래스를 이용해 중복 로직 관리
