+++ src/main/java/step3/view/InputView.java
@@ -11,6 +11,7 @@ public interface InputView {
 
     PositiveNumber tryCount();
 
+    @Deprecated
     PositiveNumber carCount();
 
     Set<CarName> carNames();
