+++ src/test/java/step4/mock/TestInputView.java
@@ -1,0 +1,17 @@
+package step4.mock;
+
+import step3.view.AbstractInputView;
+
+public class TestInputView extends AbstractInputView {
+
+    private final String input;
+
+    public TestInputView(String input) {
+        this.input = input;
+    }
+
+    @Override
+    public String read() {
+        return input.trim();
+    }
+}
